+++ index.html
@@ -230,5 +230,7 @@
     </div>
   </template>
 <script src="js/main.js"></script>
+<script type="module" src="js/util.js"></script>
+<script type="module" src="js/data.js"></script>
 </body>
 </html>
