+++ js/data.js
@@ -1,0 +1,25 @@
+const SIMILAR_ADVERT_COUNT = 10;
+
+const MAX_PRICE = 10000;
+const MAX_ROOMS = 10000;
+const MAX_GUESTS = 10000;
+
+const LAT_MIN = 35.65000;
+const LAT_MAX = 35.70000;
+const LNG_MIN = 139.70000;
+const LNG_MAX = 139.80000;
+
+const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const PHOTOS = ['https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+const CHECKIN_TIMES = ['12:00', '13:00', '14:00'];
+const CHECKOUT_TIMES = ['12:00', '13:00', '14:00'];
+const TITLES = ['Квартира', 'Отель', 'Комната', 'Кемпинг', 'Трейлер'];
+const TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+
+
+export {SIMILAR_ADVERT_COUNT, MAX_PRICE, MAX_ROOMS, MAX_GUESTS, LAT_MIN, LAT_MAX, LNG_MIN, LNG_MAX, FEATURES,
+  PHOTOS, CHECKIN_TIMES, CHECKOUT_TIMES, TITLES, TYPES};
